cmount: add support for fuse3
